Fixed - Multimap cluster compatibility. #1286
