+++ redisson/src/main/java/org/redisson/RedissonListMultimap.java
@@ -62,12 +62,13 @@ public class RedissonListMultimap<K, V> extends RedissonMultimap<K, V> implement
                 "local size = 0; " +
                 "for i, v in ipairs(keys) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " +
+                        "local name = ARGV[1] .. v; " +
                         "size = size + redis.call('llen', name); " +
                     "end;" +
                 "end; " +
                 "return size; ",
-                Arrays.<Object>asList(getName()));
+                Arrays.<Object>asList(getName()), 
+                prefix);
     }
     
     @Override
@@ -87,7 +88,7 @@ public class RedissonListMultimap<K, V> extends RedissonMultimap<K, V> implement
                 "local keys = redis.call('hgetall', KEYS[1]); " +
                 "for i, v in ipairs(keys) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " +
+                        "local name = ARGV[2] .. v; " +
 
                         "local items = redis.call('lrange', name, 0, -1) " +
                         "for i=1,#items do " +
@@ -98,7 +99,8 @@ public class RedissonListMultimap<K, V> extends RedissonMultimap<K, V> implement
                     "end;" +
                 "end; " +
                 "return 0; ",
-                Arrays.<Object>asList(getName()), valueState);
+                Arrays.<Object>asList(getName()), 
+                valueState, prefix);
     }
 
     @Override
