+++ redisson/src/main/java/org/redisson/RedissonMultimap.java
@@ -15,7 +15,6 @@
  */
 package org.redisson;
 
-import java.net.InetSocketAddress;
 import java.util.AbstractCollection;
 import java.util.AbstractSet;
 import java.util.ArrayList;
@@ -58,15 +57,18 @@ import io.netty.buffer.ByteBuf;
 public abstract class RedissonMultimap<K, V> extends RedissonExpirable implements RMultimap<K, V> {
 
     private final UUID id;
+    final String prefix;
     
     RedissonMultimap(UUID id, CommandAsyncExecutor connectionManager, String name) {
         super(connectionManager, name);
         this.id = id;
+        prefix = suffixName(getName(), "");
     }
 
     RedissonMultimap(UUID id, Codec codec, CommandAsyncExecutor connectionManager, String name) {
         super(codec, connectionManager, name);
         this.id = id;
+        prefix = suffixName(getName(), "");
     }
 
     @Override
@@ -139,7 +141,7 @@ public abstract class RedissonMultimap<K, V> extends RedissonExpirable implement
     }
 
     String getValuesName(String hash) {
-        return "{" + getName() + "}:" + hash;
+        return suffixName(getName(), hash);
     }
 
     @Override
@@ -239,7 +241,7 @@ public abstract class RedissonMultimap<K, V> extends RedissonExpirable implement
                 "local keys = {KEYS[1]}; " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[1] .. v; " + 
                         "table.insert(keys, name); " +
                     "end;" +
                 "end; " +
@@ -249,7 +251,7 @@ public abstract class RedissonMultimap<K, V> extends RedissonExpirable implement
                     + "n = n + redis.call('del', unpack(keys, i, math.min(i+4999, table.getn(keys)))) "
                 + "end; "
                 + "return n;",
-                Arrays.<Object>asList(getName()));
+                Arrays.<Object>asList(getName()), prefix);
     }
 
     @Override
@@ -258,12 +260,13 @@ public abstract class RedissonMultimap<K, V> extends RedissonExpirable implement
                 "local entries = redis.call('hgetall', KEYS[1]); " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[2] .. v; " + 
                         "redis.call('pexpire', name, ARGV[1]); " +
                     "end;" +
                 "end; " +
                 "return redis.call('pexpire', KEYS[1], ARGV[1]); ",
-                Arrays.<Object>asList(getName()), timeUnit.toMillis(timeToLive));
+                Arrays.<Object>asList(getName()), 
+                timeUnit.toMillis(timeToLive), prefix);
     }
 
     @Override
@@ -272,12 +275,13 @@ public abstract class RedissonMultimap<K, V> extends RedissonExpirable implement
                 "local entries = redis.call('hgetall', KEYS[1]); " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[2] .. v; " + 
                         "redis.call('pexpireat', name, ARGV[1]); " +
                     "end;" +
                 "end; " +
                 "return redis.call('pexpireat', KEYS[1], ARGV[1]); ",
-                Arrays.<Object>asList(getName()), timestamp);
+                Arrays.<Object>asList(getName()), 
+                timestamp, prefix);
     }
 
     @Override
@@ -286,12 +290,13 @@ public abstract class RedissonMultimap<K, V> extends RedissonExpirable implement
                 "local entries = redis.call('hgetall', KEYS[1]); " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[1] .. v; " + 
                         "redis.call('persist', name); " +
                     "end;" +
                 "end; " +
                 "return redis.call('persist', KEYS[1]); ",
-                Arrays.<Object>asList(getName()));
+                Arrays.<Object>asList(getName()),
+                prefix);
     }
     
     @Override
