+++ redisson/src/main/java/org/redisson/RedissonMultimapCache.java
@@ -24,16 +24,24 @@ import org.redisson.client.codec.LongCodec;
 import org.redisson.client.protocol.RedisCommands;
 import org.redisson.command.CommandAsyncExecutor;
 
+/**
+ * 
+ * @author Nikita Koksharov
+ *
+ * @param <K> key type
+ */
 public class RedissonMultimapCache<K> {
 
     private final CommandAsyncExecutor commandExecutor;
     private final RObject object;
     private final String timeoutSetName;
+    private final String prefix;
     
-    public RedissonMultimapCache(CommandAsyncExecutor commandExecutor, RObject object, String timeoutSetName) {
+    public RedissonMultimapCache(CommandAsyncExecutor commandExecutor, RObject object, String timeoutSetName, String prefix) {
         this.commandExecutor = commandExecutor;
         this.object = object;
         this.timeoutSetName = timeoutSetName;
+        this.prefix = prefix;
     }
 
     public RFuture<Boolean> expireKeyAsync(K key, long timeToLive, TimeUnit timeUnit) {
@@ -60,7 +68,7 @@ public class RedissonMultimapCache<K> {
                 "local keys = {KEYS[1], KEYS[2]}; " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[1] .. v; " + 
                         "table.insert(keys, name); " +
                     "end;" +
                 "end; " +
@@ -70,7 +78,8 @@ public class RedissonMultimapCache<K> {
                     + "n = n + redis.call('del', unpack(keys, i, math.min(i+4999, table.getn(keys)))) "
                 + "end; "
                 + "return n;",
-                Arrays.<Object>asList(object.getName(), timeoutSetName));
+                Arrays.<Object>asList(object.getName(), timeoutSetName), 
+                prefix);
     }
 
     public RFuture<Boolean> expireAsync(long timeToLive, TimeUnit timeUnit) {
@@ -79,13 +88,14 @@ public class RedissonMultimapCache<K> {
                 "local entries = redis.call('hgetall', KEYS[1]); " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[2] .. v; " + 
                         "redis.call('pexpire', name, ARGV[1]); " +
                     "end;" +
                 "end; " +
                 "redis.call('pexpire', KEYS[2], ARGV[1]); " +
                 "return redis.call('pexpire', KEYS[1], ARGV[1]); ",
-                Arrays.<Object>asList(object.getName(), timeoutSetName), timeUnit.toMillis(timeToLive));
+                Arrays.<Object>asList(object.getName(), timeoutSetName), 
+                timeUnit.toMillis(timeToLive), prefix);
     }
 
     public RFuture<Boolean> expireAtAsync(long timestamp) {
@@ -94,13 +104,14 @@ public class RedissonMultimapCache<K> {
                 "local entries = redis.call('hgetall', KEYS[1]); " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[2] .. v; " + 
                         "redis.call('pexpireat', name, ARGV[1]); " +
                     "end;" +
                 "end; " +
                 "redis.call('pexpireat', KEYS[2], ARGV[1]); " +
                 "return redis.call('pexpireat', KEYS[1], ARGV[1]); ",
-                Arrays.<Object>asList(object.getName(), timeoutSetName), timestamp);
+                Arrays.<Object>asList(object.getName(), timeoutSetName),
+                timestamp, prefix);
     }
 
     public RFuture<Boolean> clearExpireAsync() {
@@ -109,13 +120,14 @@ public class RedissonMultimapCache<K> {
                 "local entries = redis.call('hgetall', KEYS[1]); " +
                 "for i, v in ipairs(entries) do " +
                     "if i % 2 == 0 then " +
-                        "local name = '{' .. KEYS[1] .. '}:' .. v; " + 
+                        "local name = ARGV[1] .. v; " + 
                         "redis.call('persist', name); " +
                     "end;" +
                 "end; " +
                 "redis.call('persist', KEYS[2]); " +
                 "return redis.call('persist', KEYS[1]); ",
-                Arrays.<Object>asList(object.getName(), timeoutSetName));
+                Arrays.<Object>asList(object.getName(), timeoutSetName),
+                prefix);
     }
